fix: address mock usage review feedback
