+++ pkg/services/workers/internal/services/runners/wire/mock_registry_test.go
@@ -9,6 +9,7 @@ import (
 
 	"github.com/portpowered/infinite-you/pkg/services/work"
 	"github.com/portpowered/infinite-you/pkg/services/workers"
+	workerexecution "github.com/portpowered/infinite-you/pkg/services/workers/internal/execution"
 	"github.com/portpowered/infinite-you/pkg/services/workers/internal/services/runners"
 	"github.com/portpowered/infinite-you/pkg/services/workers/internal/services/runners/internal/testkit"
 	workerprocess "github.com/portpowered/infinite-you/pkg/services/workers/internal/services/runners/process"
@@ -40,6 +41,11 @@ func TestNewMockProductionRegistryIsInertAndResolvesWithoutExecution(t *testing.
 }
 
 func TestMockRunnerThroughProductionRegistryConformsToCommonContract(t *testing.T) {
+	inputTokens, outputTokens := int64(10), int64(5)
+	usage := &workers.MockWorkerUsageConfig{
+		Provider: "codex", Model: "gpt-5-codex",
+		InputTokens: &inputTokens, OutputTokens: &outputTokens,
+	}
 	registry := newExplicitMockRegistry(t, &workers.MockWorkersConfig{
 		MockWorkers: []workers.MockWorkerConfig{{
 			WorkerName: "writer",
@@ -47,6 +53,7 @@ func TestMockRunnerThroughProductionRegistryConformsToCommonContract(t *testing.
 		}, {
 			WorkerName: "failing",
 			RunType:    workers.MockWorkerRunTypeReject,
+			Usage:      usage,
 		}},
 	}, nil)
 
@@ -65,6 +72,24 @@ func TestMockRunnerThroughProductionRegistryConformsToCommonContract(t *testing.
 		workers.RunnerOptionalCapabilityImageInput,
 	}
 	failure := mockRequest("failing")
+	capture := &workerexecution.MockWorkerUsageCapture{}
+	failureResult, failureErr := registry.Execute(
+		workerexecution.WithMockWorkerUsageCapture(t.Context(), capture),
+		runners.ExecuteRequest{Identity: runners.MockIdentity, Attempt: failure},
+	)
+	if failureErr == nil {
+		t.Fatal("reject Execute() error = nil, want mock rejection")
+	}
+	if got := capture.Usage(); got == nil || got.Provider != usage.Provider || got.Model != usage.Model {
+		t.Fatalf("captured reject usage = %#v, want declared provider/model", got)
+	}
+	if failureResult.Diagnostics == nil || failureResult.Diagnostics.Provider == nil ||
+		failureResult.Diagnostics.Provider.Provider != usage.Provider ||
+		failureResult.Diagnostics.Provider.Model != usage.Model ||
+		failureResult.Diagnostics.Provider.ResponseMetadata[workers.ProviderResponseMetadataInputTokens] != "10" ||
+		failureResult.Diagnostics.Provider.ResponseMetadata[workers.ProviderResponseMetadataOutputTokens] != "5" {
+		t.Fatalf("reject diagnostics = %#v, want declared provider usage", failureResult.Diagnostics)
+	}
 
 	testkit.RunService(t, testkit.ServiceSubject{
 		Service:            registry,
